requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
numpy==1.20.3
pandas==1.2.4
plotly==4.14.3
protobuf==3.17.1
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
six==1.16.0
Werkzeug==2.0.1
